# networkTopology: optimise replicaMap slightly