```
name                   old time/op    new time/op    delta
Topology_ReplicaMap-8    83.5ms ± 1%    54.0ms ± 1%  -35.29%  (p=0.000 n=8+9)

name                   old alloc/op   new alloc/op   delta
Topology_ReplicaMap-8    5.73MB ± 0%    0.83MB ± 0%  -85.44%  (p=0.000 n=9+8)

name                   old allocs/op  new allocs/op  delta
Topology_ReplicaMap-8     89.2k ± 0%      6.9k ± 0%  -92.22%  (p=0.000 n=10+10)
```